+++ README.md
@@ -1,2 +1,2 @@
 # Personal Website
-My (personal website)[www.nidhiswamy.com] built using Next.js, App Router, and Tailwind CSS. Deployed using Vercel.
+My personal website [www.nidhiswamy.com] built using Next.js, App Router, and Tailwind CSS. Deployed using Vercel.
