+++ src/app/ui/navbar.tsx
@@ -1,4 +1,4 @@
-import NavLinks from '@/app/ui/nav-links';
+import NavLinks from '@/app/ui/NavLinks';
 import Logo from '@/app/ui/Logo';
 
 export default function NavBar() {
